requirements.txt
flask
flask-sqlalchemy
pytest
SQLAlchemy
werkzeug
PyJWT